Adding loading animation, fixing some queries
