+++ src/app/layout.tsx
@@ -15,7 +15,7 @@ const fontSans = FontSans({
   variable: "--font-sans",
 })
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
