+++ src/app/prompt/[[...slug]]/prompt-page.tsx
@@ -1,0 +1,86 @@
+"use client"
+import React, { useState } from 'react'
+import { Textarea } from '@/components/ui/textarea'
+import { Button } from '@/components/ui/button'
+import { MessageProps } from '@/types'
+import ChatBox from '@/components/chat-box'
+import { answerQuestions } from '@/lib/utils'
+import { UserProfileProps } from '@/types'
+import { useAuth } from "@clerk/nextjs"
+
+
+type Props = {
+    user: UserProfileProps | null
+}
+
+const PromptPage = ({ user }: Props) => {
+    const [data, setData] = useState<MessageProps[]>([])
+    const [prompt, setPrompt] = useState<string>("")
+    const [isLoading, setIsLoading] = useState<boolean>(false)
+
+    const handleSendPrompt = (e:React.FormEvent<HTMLFormElement>) => {
+        e.preventDefault()
+        const newMessage = {
+            type: "request",
+            message: prompt
+        }
+        const newData = [...data, newMessage]
+        setData(newData)
+        setIsLoading(true)
+        handleGetResponse().then(res => {
+            setData([...newData, res])
+            setIsLoading(false)
+        })
+        setPrompt("")
+    }
+
+    const handleKeyPressDown = (e:React.KeyboardEvent<HTMLFormElement>) => {
+        if (e.key === 'Enter' && !e.shiftKey) {
+            handleSendPrompt(e)
+        }
+    }
+
+    const handleGetResponse = async () => {
+        const data = await answerQuestions(prompt)
+        // async
+        const newResponse = {
+            type: "response",
+            message: data
+        }
+        return newResponse      
+    }
+
+  return (
+    <div className='flex flex-col w-full p-5 h-screen'>
+        <div className="w-full flex-1 overflow-y-auto mb-5 py-5 px-[100px]">
+            {data.length === 0 && <div className="flex flex-col justify-center h-full max-w-[900px] m-auto">
+                <div className="bg-gradient-to-r from-blue-500 to-teal-300 bg-clip-text text-transparent animate-slide-in delay-300">
+                    <h1 className='text-6xl font-medium py-3'>Welcome, {user &&  user.username }</h1>
+                </div>
+                <h1 className='text-6xl font-medium py-3 bg-gradient-to-r from-neutral-500 to-sky-700 bg-clip-text text-transparent animate-slide-in delay-300'>Ready to learn something new?</h1>
+            </div>}
+            <div className="flex flex-col m-auto max-w-[900px] pt-10">
+                { data.map((item, i) => (
+                    <ChatBox variant={item.type} message={item.message} key={i}/>
+                )) }
+                { isLoading && 
+                <div className="flex w-full text-blue-700 items-center text-sm justify-center">
+                    <svg aria-hidden="true" role="status" className="inline w-4 h-4 me-3 text-gray-200 animate-spin dark:text-gray-600" viewBox="0 0 100 101" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <path d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z" fill="currentColor"/>
+                    <path d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z" fill="#1C64F2"/>
+                    </svg>
+                    Processing...
+                </div>}
+            </div>
+        </div>
+        <div className="flex justify-center items-center">
+            <form action="submit" onSubmit={handleSendPrompt} onKeyDown={handleKeyPressDown} className='flex w-full max-w-[900px] items-center gap-x-4'>
+                <Textarea value={prompt} onChange={(e) => setPrompt(e.target.value)} placeholder='Enter a prompt...' className='resize-none w-full'/>
+                <Button disabled={prompt.length === 0}>Send</Button>
+            </form>
+        </div>
+    </div>
+  )
+}
+
+export default PromptPage
