+++ src/middleware.ts
@@ -1,11 +1,12 @@
-import { clerkMiddleware, createRouteMatcher } from "@clerk/nextjs/server";
+import { clerkMiddleware, createRouteMatcher, clerkClient } from "@clerk/nextjs/server";
+
 
 
 const isProtectedRoute = createRouteMatcher([
     '/dashboard/(.*)', '/prompt(.*)'
 ])
 
-export default clerkMiddleware((auth, req) => {
+export default clerkMiddleware(async (auth, req) => {
     if (isProtectedRoute(req)) {
         auth().protect();
     }
